refactor(model-list): own pricing model types
