+++ e2e/modelListCommonFlows.spec.ts
@@ -4,8 +4,8 @@ import { OPTIONS_PAGE_PATH } from "~/constants/extensionPages"
 import { MENU_ITEM_IDS } from "~/constants/optionsMenuIds"
 import { API_CREDENTIAL_PROFILES_TEST_IDS } from "~/features/ApiCredentialProfiles/testIds"
 import { MODEL_LIST_TEST_IDS } from "~/features/ModelList/testIds"
-import type { ModelPricing } from "~/services/apiService/common/type"
 import { STORAGE_KEYS } from "~/services/core/storageKeys"
+import type { ModelPricing } from "~/services/modelList/pricingModel"
 import { expect, test } from "~~/e2e/fixtures/extensionTest"
 import { verifyAccountModelCatalogUsage } from "~~/e2e/scenarios/accountUsage"
 import {
