+++ src/services/apiService/newApiFamily/modelPricing.ts
@@ -1,10 +1,8 @@
 import { SITE_TYPES, type AccountSiteType } from "~/constants/siteType"
 import * as commonModelPricing from "~/services/apiService/common"
-import type {
-  ApiServiceRequest,
-  PricingResponse,
-} from "~/services/apiService/common/type"
+import { type ApiServiceRequest } from "~/services/apiService/common/type"
 import * as oneHubModelPricing from "~/services/apiService/oneHub"
+import type { PricingResponse } from "~/services/modelList/pricingModel"
 
 interface ModelPricingImplementation {
   fetchModelPricing: (request: ApiServiceRequest) => Promise<PricingResponse>
