+++ tests/features/ModelList/components/ModelItemPricing.test.tsx
@@ -10,7 +10,7 @@ import {
   MODEL_PRICE_PRECISION_KINDS,
   MODEL_PRICE_SOURCE_KINDS,
   MODEL_UNAVAILABLE_PRICE_REASONS,
-} from "~/services/apiService/common/type"
+} from "~/services/modelList/pricingModel"
 
 const { formatPriceCompactMock, isTokenBillingTypeMock } = vi.hoisted(() => ({
   formatPriceCompactMock: vi.fn(
