+++ tests/services/modelList/accountSources/tokenScopedFallback.routes.test.ts
@@ -13,7 +13,7 @@ import {
   MODEL_PRICE_PRECISION_KINDS,
   MODEL_PRICE_SOURCE_KINDS,
   MODEL_UNAVAILABLE_PRICE_REASONS,
-} from "~/services/apiService/common/type"
+} from "~/services/modelList/pricingModel"
 import { AuthTypeEnum } from "~/types"
 
 const {
