+++ tests/services/modelList/pricingResponse.test.ts
@@ -6,7 +6,7 @@ import {
   MODEL_PRICE_PRECISION_KINDS,
   MODEL_PRICE_SOURCE_KINDS,
   MODEL_UNAVAILABLE_PRICE_REASONS,
-} from "~/services/apiService/common/type"
+} from "~/services/modelList/pricingModel"
 import {
   buildModelListCatalogPricingResponse,
   normalizeModelListModelIds,
